requirements.txt
# =============================================================================
# PariShiksha — Dependencies (Wk9 + Wk10)
# =============================================================================

# --- Wk10 Core Dependencies ---
tiktoken>=0.5.0               # Token counting for chunking (cl100k_base)
chromadb>=0.4.0                # Persistent vector database
openai>=1.6.0                  # OpenAI embeddings (text-embedding-3-small)
anthropic>=0.39.0              # Anthropic Claude Haiku generation

# --- PDF Extraction & Text Processing ---
PyMuPDF>=1.23.0                # fitz — fast PDF text extraction
pdfplumber>=0.10.0             # Alternative PDF extractor
pytesseract>=0.3.10            # OCR fallback

# --- NLP & Tokenization ---
transformers>=4.36.0           # Hugging Face tokenizers/models
tokenizers>=0.15.0             # Fast tokenizer library
sentencepiece>=0.1.99          # SentencePiece tokenizer (T5)
nltk>=3.8.1                    # Sentence tokenization
spacy>=3.7.0                   # Sentence segmentation

# --- Retrieval ---
rank-bm25>=0.2.2               # BM25 keyword ranking
sentence-transformers>=2.2.2   # SBERT dense embeddings
scikit-learn>=1.3.0            # TF-IDF, metrics
numpy>=1.24.0
scipy>=1.11.0

# --- Generation ---
google-generativeai>=0.3.0     # Gemini API (Wk9 legacy)

# --- Evaluation ---
rouge-score>=0.1.2             # ROUGE metrics
bert-score>=0.3.13             # BERTScore

# --- Data & Visualization ---
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
tabulate>=0.9.0

# --- Utilities ---
python-dotenv>=1.0.0           # .env file loading
tqdm>=4.66.0                   # Progress bars
pyyaml>=6.0
loguru>=0.7.0                  # Structured logging

# --- Testing ---
pytest>=7.4.0
pytest-cov>=4.1.0
